move testcafe stuff in own package with own node_modules folder
