+++ package.json
@@ -9,9 +9,7 @@
     "test-misc": "node ./bin/test_package_lock_version",
     "test-log-conventions": "node ./bin/log-conventions",
     "test-unit": "env TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\" }' mocha -r ts-node/register 'test/mocha/**/*.ts'",
-    "test-e2e-ci": "env NODE_ENV=test testcafe --skip-js-errors \"electron:./\" \"./test/testcafe/test.js\" && npm run test-e2e-msglist",
-    "test-e2e": "env NODE_ENV=test testcafe --skip-js-errors \"electron:./\" \"./test/testcafe/test.js\" && npm run test-e2e-msglist",
-    "test-e2e-msglist": " TEST_DIR=\".test_tmp_data\" NODE_ENV=test npx testcafe --skip-js-errors \"electron:./\" \"./test/testcafe/messagelist_tests.ts\" --compilerOptions.typescript.configPath=test/testcafe/tsconfig.json",
+    "test-e2e": "cd e2e_tests && npm i && npm run start-test",
     "check-types": "tsc -b src/check-all-tsconfig.json",
     "check-formatting": "prettier --check \"scss/**/*.scss\" && prettier --check \"src/**/*.{js,ts,tsx,json}\" && prettier --check \"**/*.md\"",
     "fix-formatting": "prettier --write \"scss/**/*.scss\" && prettier --write \"src/**/*.{js,ts,tsx,json}\" && prettier --write \"**/*.md\"",
@@ -140,9 +138,6 @@
     "node-fetch": "^2.6.7",
     "prettier": "2.2.1",
     "sass": "^1.26.5",
-    "testcafe": "1.18.5",
-    "testcafe-browser-provider-electron": "0.0.18",
-    "testcafe-react-selectors": "^4.1.5",
     "ts-node": "^10.4.0",
     "typescript": "^4.4.4",
     "walk": "^2.3.14",
